+++ src/components/Avatar/Avatar.styled.js
