+++ src/components/DropDown/DropDown.styled.js
@@ -1,0 +1,11 @@
+import styled from '@emotion/styled';
+import colors from '../../styles/colors';
+
+const BaseDropDown = styled.div`
+  border-radius: 0.5rem;
+  border: 0.06rem solid ${colors['--Gray-300']};
+  background-color: ${colors['--White']};
+  box-shadow: 0 0.125rem 0.75rem 0 rgba(0, 0, 0, 0.08);
+`;
+
+export default BaseDropDown;
