+++ src/pages/RollingPaperPage/RollingPaperPage.jsx
@@ -2,11 +2,11 @@ import { useEffect } from 'react';
 import { useInView } from 'react-intersection-observer';
 import { useLocation, useParams } from 'react-router-dom';
 import { Helmet } from 'react-helmet-async';
-import CardList from '../../components/RollingPaper/MessageCardList/MessageCardList';
-import MessageModal from '../../components/RollingPaper/MessageModal/MessageModal';
+import CardList from '../../components/MessageCardList/MessageCardList';
+import MessageModal from '../../components/MessageModal/MessageModal';
 import useModal from '../../utils/hooks/useModal';
 import useGetMessages from '../../utils/hooks/useGetMessages';
-import Header from '../../components/RollingPaper/HeaderService/HeaderService';
+import Header from '../../components/HeaderService/HeaderService';
 import useDeleteRollingPaper from '../../utils/hooks/useDeleteRollingPaper';
 import { Primary40Button } from '../../components/common/Button/Button';
 import { isLogin } from '../../utils/helpers/validator';
